Fraus/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
motor==3.7.1
pydantic==2.11.7
python-dotenv==1.0.1
httpx==0.28.1
